+++ app/dashboard/views.py
@@ -136,6 +136,12 @@ def result_change_project(request):
 def scenario_visualize_results(
     request, proj_id=None, scen_id=None, step_id=5, max_step=MAX_STEP
 ):
+    page_information = mark_safe(
+        "Please be aware that the tool is still under development and results are subject to change. "
+        "For further information on the applied methods please take a look at the "
+        "<a href=https://open-plan-documentation.readthedocs.io/en/latest/getting_started/basic_structure.html target='_blank'>documentation</a>."
+    )
+
     request.session[COMPARE_VIEW] = False
 
     user_projects = fetch_user_projects(request.user)
@@ -184,6 +190,7 @@ def scenario_visualize_results(
                 request,
                 "report/single_scenario.html",
                 {
+                    "page_information": page_information,
                     "project_list": user_projects,
                     "proj_id": proj_id,
                     "scen_id": scen_id,
@@ -228,6 +235,7 @@ def scenario_visualize_results(
                     request,
                     "report/single_scenario.html",
                     {
+                        "page_information": page_information,
                         "scen_id": scen_id,
                         "scenario": scenario,
                         "step_list": STEP_LIST,
@@ -266,6 +274,11 @@ def scenario_visualize_results(
 def project_compare_results(request, proj_id, step_id=5, max_step=MAX_STEP):
     request.session[COMPARE_VIEW] = True
     user_projects = fetch_user_projects(request.user)
+    page_information = mark_safe(
+        "Please be aware that the tool is still under development and results are subject to change. "
+        "For further information on the applied methods please take a look at the "
+        "<a href=https://open-plan-documentation.readthedocs.io/en/latest/getting_started/basic_structure.html target='_blank'>documentation</a>."
+    )
 
     project = get_object_or_404(Project, id=proj_id)
     if (project.user != request.user) and (
@@ -287,6 +300,7 @@ def project_compare_results(request, proj_id, step_id=5, max_step=MAX_STEP):
         "report/compare_scenario.html",
         {
             "scen_id": None,
+            "page_information": page_information,
             "step_list": STEP_LIST,
             "max_step": max_step,
             "step_id": step_id,
@@ -687,10 +701,15 @@ def request_kpi_table(request, proj_id=None):
                         break
 
                     if param["unit"] == "%":
-                        val = round(float(val) * 100, 2)
-                    values.append(beautify_number(val, 2))
+                        values.append(beautify_number(val, 2))
+                    else:
+                        values.append(beautify_number(val, 0))
 
                 if missing:
+                    continue
+
+                # skip KPI for carrier demand if all scenario values are 0
+                if "demand" in param["id"] and all(val == "0" for val in values):
                     continue
 
                 param["scen_values"] = values
@@ -789,27 +808,35 @@ def view_asset_parameters(request, scen_id, asset_type_name, asset_uuid):
         ess_discharging_power_asset = ess_asset_children.get(
             asset_type__asset_type="discharging_power"
         )
+
+        initial = {
+            "name": existing_ess_asset.name,
+            "installed_capacity": ess_capacity_asset.installed_capacity,
+            "age_installed": ess_capacity_asset.age_installed,
+            "capex_fix": ess_capacity_asset.capex_fix,
+            "capex_var": ess_capacity_asset.capex_var,
+            "opex_fix": ess_capacity_asset.opex_fix,
+            "opex_var": ess_capacity_asset.opex_var,
+            "lifetime": ess_capacity_asset.lifetime,
+            "crate": ess_capacity_asset.crate,
+            "efficiency": ess_capacity_asset.efficiency,
+            "dispatchable": ess_capacity_asset.dispatchable,
+            "optimize_cap": ess_capacity_asset.optimize_cap,
+            "soc_max": ess_capacity_asset.soc_max,
+            "soc_min": ess_capacity_asset.soc_min,
+        }
+
+        # Add thermal loss rate fields to initial if hess
+        if asset_type_name == "hess":
+            for field in [
+                "thermal_loss_rate",
+                "fixed_thermal_losses_relative",
+                "fixed_thermal_losses_absolute",
+            ]:
+                initial[field] = getattr(ess_capacity_asset, field)
+
         # also get all child assets
-        form = StorageForm(
-            asset_type=asset_type_name,
-            view_only=True,
-            initial={
-                "name": existing_ess_asset.name,
-                "installed_capacity": ess_capacity_asset.installed_capacity,
-                "age_installed": ess_capacity_asset.age_installed,
-                "capex_fix": ess_capacity_asset.capex_fix,
-                "capex_var": ess_capacity_asset.capex_var,
-                "opex_fix": ess_capacity_asset.opex_fix,
-                "opex_var": ess_capacity_asset.opex_var,
-                "lifetime": ess_capacity_asset.lifetime,
-                "crate": ess_capacity_asset.crate,
-                "efficiency": ess_capacity_asset.efficiency,
-                "dispatchable": ess_capacity_asset.dispatchable,
-                "optimize_cap": ess_capacity_asset.optimize_cap,
-                "soc_max": ess_capacity_asset.soc_max,
-                "soc_min": ess_capacity_asset.soc_min,
-            },
-        )
+        form = StorageForm(asset_type=asset_type_name, view_only=True, initial=initial)
         optimized_cap = ess_capacity_asset.optimize_cap
         existing_asset = existing_ess_asset